% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoniche_model.R
\name{demoniche_model}
\alias{demoniche_model}
\title{Run spatially explicit demographic model}
\usage{
demoniche_model(modelname, Niche, Dispersal, repetitions = 1000, foldername)
}
\arguments{
\item{modelname}{Name of \code{list} object containing model parameters}

\item{Niche}{TRUE/FALSE that specifies whether to run the simulation accounting for effects of Niche values}

\item{Dispersal}{TRUE/FALSE that specifies whether to allow long-distance dispersal}

\item{repetitions}{Number of replicate simulations to carry out (default = 1000)}

\item{foldername}{Path name of folder where results of simulations will be written to.}
}
\value{
Something very special
}
\description{
Run spatially explicit demographic model
}
